Close #432 - syntax should have cats (effectie.cats.syntax and effectie.syntax => effectie.syntax)
