+++ modules/effectie-cats-effect/shared/src/test/scala-3/effectie/cats/syntax/errorSpec.scala
@@ -1,12 +1,12 @@
-package effectie.cats.syntax
+package effectie.syntax
 
 import cats.*
 import cats.data.EitherT
 import cats.effect.*
 import cats.syntax.all.*
 import effectie.SomeControlThrowable
-import effectie.cats.syntax.fx.*
-import effectie.cats.syntax.error.*
+import effectie.syntax.fx.*
+import effectie.syntax.error.*
 import effectie.core.{Fx, FxCtor}
 import effectie.testing.types.SomeError
 import extras.concurrent.testing.ConcurrentSupport
