--- modules/effectie-cats/shared/src/main/scala-2/effectie/cats/syntax/all.scala
@@ -1,7 +1,0 @@
-package effectie.cats.syntax
-
-/** @author Kevin Lee
-  * @since 2021-05-18
-  */
-trait all extends fx with error with console
-object all extends all
