+++ modules/effectie-cats/shared/src/main/scala-2/effectie/syntax/all.scala
@@ -1,7 +1,7 @@
-package effectie.cats.syntax
+package effectie.syntax
 
 /** @author Kevin Lee
-  * @since 2021-10-23
+  * @since 2021-05-18
   */
 trait all extends fx with error with console
 object all extends all
