+++ modules/effectie-cats/shared/src/main/scala-2/effectie/syntax/fx.scala
@@ -3,10 +3,11 @@ package effectie.syntax
 import effectie.core.FxCtor
 
 /** @author Kevin Lee
-  * @since 2022-01-17
+  * @since 2022-09-08
   */
 trait fx {
-  import effectie.syntax.fx.*
+
+  import fx._
 
   def effectOf[F[*]]: CurriedEffectOf[F] = new CurriedEffectOf[F]
 
@@ -14,39 +15,38 @@ trait fx {
 
   def pureOrError[F[*]]: CurriedEffectOfPureOrError[F] = new CurriedEffectOfPureOrError[F]
 
-  inline final def unitOf[F[*]: FxCtor]: F[Unit] = FxCtor[F].unitOf
+  def unitOf[F[*]: FxCtor]: F[Unit] = FxCtor[F].unitOf
 
   def errorOf[F[*]]: CurriedErrorOf[F] = new CurriedErrorOf[F]
 
 }
-
 object fx extends fx {
 
   private[fx] final class CurriedEffectOf[F[*]](
     private val dummy: Boolean = true
   ) extends AnyVal {
-    def apply[A](a: => A)(using EF: FxCtor[F]): F[A] =
+    def apply[A](a: => A)(implicit EF: FxCtor[F]): F[A] =
       FxCtor[F].effectOf(a)
   }
 
   private[fx] final class CurriedEffectOfPure[F[*]](
     private val dummy: Boolean = true
   ) extends AnyVal {
-    def apply[A](a: A)(using EF: FxCtor[F]): F[A] =
+    def apply[A](a: A)(implicit EF: FxCtor[F]): F[A] =
       FxCtor[F].pureOf(a)
   }
 
   private[fx] final class CurriedEffectOfPureOrError[F[*]](
     private val dummy: Boolean = true
   ) extends AnyVal {
-    def apply[A](a: => A)(using EF: FxCtor[F]): F[A] =
+    def apply[A](a: => A)(implicit EF: FxCtor[F]): F[A] =
       FxCtor[F].pureOrError(a)
   }
 
   private[fx] final class CurriedErrorOf[F[*]](
     private val dummy: Boolean = true
   ) extends AnyVal {
-    def apply[A](throwable: Throwable)(using EF: FxCtor[F]): F[A] =
+    def apply[A](throwable: Throwable)(implicit EF: FxCtor[F]): F[A] =
       FxCtor[F].errorOf(throwable)
   }
