+++ modules/effectie-cats/shared/src/main/scala-3/effectie/cats.scala
