--- modules/effectie-cats/shared/src/main/scala-3/effectie/cats/syntax/fx.scala
@@ -1,7 +1,0 @@
-package effectie.cats.syntax
-
-/** @author Kevin Lee
-  * @since 2022-09-08
-  */
-trait fx extends effectie.syntax.fx
-object fx extends fx
