+++ modules/effectie-cats/shared/src/main/scala-3/effectie/syntax/all.scala
@@ -1,7 +1,7 @@
 package effectie.syntax
 
 /** @author Kevin Lee
-  * @since 2022-01-17
+  * @since 2021-10-23
   */
 trait all extends fx with error with console
 object all extends all
