+++ modules/effectie-cats/shared/src/test/scala-2/effectie/syntax/errorSpec.scala
