+++ modules/effectie-cats/shared/src/test/scala-3/effectie/syntax/errorSpec.scala
