--- modules/effectie-core/shared/src/main/scala-2/effectie/syntax/all.scala
@@ -1,7 +1,0 @@
-package effectie.syntax
-
-/** @author Kevin Lee
-  * @since 2022-01-18
-  */
-trait all extends fx with error with console
-object all extends all
